Build: Enable FORCE_COLOR=1 in CI
This is something we lost in the transition from Travis, where procs
had a more consistently detectable TTY. There's a number of ways to
make this work by magic, but it actually makes sense for those to
not be TTYs as far as I'm concerned. Set it explicitly which is what
we're trying to do.