Core: Fix internal circular dependency from test.js to core.js
This is an ancient left-over from when the async test API consisted
of `QUnit.stop` and `QUnit.start` (as opposed to `assert.async()` and
the built-in test promise await), where `QUnit.start` was also at the
same time responsible for the initial starting of executing the test
run for cases where autostart=false (aka "begin").

Part of this legacy remains at https://github.com/qunitjs/qunit/issues/1084.

== Rollup, before

> (!) Circular dependencies
> src/core.js -> src/test.js -> src/core.js
> src/test.js -> src/core/processing-queue.js -> src/test.js

== Rollup, after

> (!) Circular dependency
> src/test.js -> src/core/processing-queue.js -> src/test.js