Build: Fix native use of ES6 Map
Remove global-this-polyfill, and replace it with a simpler alternative,
similer to what we already rely on in lib/promise-polyfill.js.

Then, with this available internally, move our es6-map.js back into
the regular build (thus effectively reverting 259f9af1bfc) where we
can access it as a regular exported value for internal use, including
the use of globalThis for accessing the native Map when available.

This resolves the canundrum we had before, where either `typeof Map`
is done within the bundle and it gets renamed by Rollup, or we do
it outside the bundle and it always finds undefined due to still being
within the outer IIFE (which was at the same time a good thing as we
don't want to leak this export to outside the QUnit scope). The fix
is obvious in retrospect, which is to access it via globalThis.

It seems we can consolidate the export.js handling for Web Workers
with the globalThis fallback (for SpiderMonkey), so let's do that.