HTML Reporter: Update fuzzysort.js to 1.2.1
Reviewed package diff via
<https://diff.intrinsic.com/fuzzysort/1.1.4/1.2.1>.

The main reason for updating is to bring in my upstream fix
(ref https://github.com/farzher/fuzzysort/pull/72/, and
ref https://github.com/farzher/fuzzysort/pull/93) to not depend
on ES6 Map being globally defined.

This unblocks the next commit, which is removing the hack for loading
es6-map.js in an unusual way. That unusual way actually caused both
our own use in test.js and fuzzysort.js to *not* see the native Map
anymore in supporting environents, which I'll fix as well.

Also, since this is a dependency that we bundle as part of the
distribution, pin its version exactly, as we do with `kleur`.