HTML Reporter: Fix retention of dropdown state over multiple searches
While it was possible to select multiple modules from the list,
whenever a new search was typed, the previous selection was lost.

This regression started in 2.11.0 with the introduction of fuzzysort.
Before that, we preserved the DOM and toggled the item's CSS display.
It changed to re-creating the HTML each time, and the function for
that was only meant for the initial render where we set checked state
based on the applied runtime configuration (QUnit.config.moduleId).

Changes:

* Keep a cache of the <li> items we create during the initial render,
  including (most importantly) its child element, the input checkbox.

* Use this cache list as the basis for fuzzysearch. Previously, we
  were using the internal objects from `QUnit.config.modules` to hang
  things like fuzzysearch's `namedPrepared` cache on. The new cache
  list has each result associated with the DOM element we need to
  render. Whenever we have results, replace the dropdown with the
  DOM elements for these results. The same DOM references remain
  re-used and thus the bug is fixed.

* Remove our custom `namePrepared` cache. Since a few versions,
  fuzzysearch has its own cache already that it uses automatically
  when you feed don't feed it already-prepared values.

* Remove toLowerCase() call. fuzzysearch does this already.

* Optimise applyUrlParams() and selectionChange() by not querying
  the DOM to rediscover our own input elements. Use our cache list
  instead.

Fixes https://github.com/qunitjs/qunit/issues/1683.