Test: Remove flaky iOS 7 from periodic full build
This particular cloud worker is tied to the deprecated simulator
cluster in BrowserStack and has been flaky for a while. Let's remove
it for now. This does not remove support for it, it just means we
lose part of the automated coverage. We still have coverage through
eslint-plugin-compat and more generally through there not being
much difference between this and other older browsers in terms of
what ECMAScript and DOM APIs we use.