Core: Fix partial Map support for IE 11 and iOS 8
Follows-up 06c39517a1, which introduced use of `new Map(iterable)`
and `Maps#keys()`, and added the relevant fallbacks, but I did not
account for the fact that:

* Map#keys was not spec'ed until ES7 and IE11 has the ES6 version.
* iOS 8 shipped an non-compliant version of ES6 Map in which
  `new Map(iterable)` was not yet supported, which was rectified
  in iOS9.

The quick commit CI job runs on IE9 which did pass before already.
The fuller pre-release CI caught the regression in IE 11 and iOS 8.