Core: Fix event "runtime" data to be rounded to milliseconds
Follows-up b70af72489, which corrected this in the UI, but this
means any plugins and integrations reporting `details.runtime` are
still affected by the needless precision.

This affects:
* `QUnit.log`, `QUnit.testDone`, `QUnit.moduleDone` and `QUnit.done`.
* `on('testEnd')`, `on('suiteEnd')`, and `on('runEnd')`.

I was initially cautious about this as I worried that perhaps
consumers would accumulate these and then become increasingly less
accurate in large test suites where perhaps most leaf nodes round to
0ms. However I have no such case in practice. Generally, for total
times the reported totals can be, should be, and are being used.

Ref https://github.com/qunitjs/qunit/issues/1678.