HTML Reporter: Optimize toolbar setup by reusing `beginDetails.modules`
`config.modules` is an undocumented internal property with lots of
other internal state in it. Relying on it here is suboptimal, and
creating a copy of it is inefficient.

Use `beginDetails.modules` instead which exists precisely for this
purpose. It was added by commit 168b048 in QUnit 1.16 and initially
used by HTML Reporter. However, commit 43a3d87 in QUnit 2.7 removed
its use again and in c494fb8af I removed the unused variables that
were left begin there.