Core: Change most internal use of objectType() to typeof operator
For cases that check for undefined, this is fully identical.

For cases that check for string, this would no longer tolerate
a String object. This is generally supported, so I've only changed
these checks in `QUnit.on()` and `assert.step()` which expect a
string literal and/or have their string compared by deepEqual later.

For cases that check for function, similarly kept those where
application code may be involved, but internal use such as the
closure passed to `assert.throws()` or `QUnit.on()` must receive
a genuine function and not some custom object that happens to
have an `obj.toString()` of `[object Function]`. Even modern
constructs like `new Proxy(function() {}, {})`, and arrow functions,
and generator functions have `typeof` evaluate to "function".