Test: Convert more ad-hoc asserts to generic tap-outputs.js cases
Various tests were still using ad-hoc asserts and `string.includes()` when
their output was actually deterministic with the current normalization for
internal frames etc, and thus can be handled the generic way instead.

I'm moving these over in preparation for
https://github.com/qunitjs/qunit/issues/1684.

Some did need normalization for `/bin` and `/src/cli`, which are hereby added.

Remove any use of `assert.pushResult({ result: false }` statements
which was only sometimes done, and is redundant anyway given that the
case of an expected-failure not happening, results in no assertions
and thus the test will already fail. This does mean that in those
cases, the output is not printed when running qunit's own tests. This
does not affect end-users of QUnit, and I'll bring it back in some
form as part of https://github.com/qunitjs/qunit/issues/1684.