HTML Reporter: Restore float-clear following e31c8d37b
I broke this element, and didn't notice it during testing because
on all the tiniest of viewports, this clearfix is actually not needed
because we use flexbox, which already makes each row in the interface
naturally float-clearing. It's only in small sub-500px height viewports
that the bug manifests.

Also add a min-height to ensure the action buttons won't fall outside
the rectangle. This is possible since the previous version because we
now use absolute position on wider viewports instead of float. This
has the benefit of ensuring the result lines and their hover/active
state are consistently full-width and continuous, but also means the
buttons aren't naturally contained within the dropdown layout.