Docs: Avoid "bad example" when "failing example" is intended
There are some pages where we use good/bad to indicate whether
the way a test is written is recommended or not.

But on deepEqual and notEqual, what we meant was actually whether
the test is passing or failing. Refer to it as such instead, this
is also consistent with what we write on other pages already.