Build: Remove iOS 8,9,10 simulators, add iOS 10 real mobile
iOS 8 is being pinned to an emulator again for some reason, and its
consistently crashing/timing out. Using the Live interface, the
controls are unresponsive.

We already have Safari 9 coverage via macOS, that'll have to suffice
and we're supposed to move away from the deprecated emulator infra
from BrowserStack anyway.

There is an iOS 10 real device available, so add that one instead.