Core: Fix inaccurate test count in reporter output after re-run
When using the HTML Reporter to re-run a specific test, the summary
would accurately report the number of assertions, but inaccurately
report the number of tests.

This was because the `Test.valid()` function, which decides whether
the test will run and is where the testId re-run filter is applied,
was called at different points with different state. Once it was called
in the constructor when we pass it to the TestReport, and then again
later at run-time to definitively decide whether to run it.

The `valid()` function isn't meant to be able to change what it
returns, but the first call could wrongly returns false because it
observes that a `config.testId` filter is active, but `this.testId`
is not yet assigned at this point in the constructor.

Fixes https://github.com/qunitjs/qunit/issues/1687.