Assert: Revert change in how deepEqual treats imposter objects
As part of https://github.com/qunitjs/qunit/pull/1700, there was an
inintended change to how we compare constructors. Instead of comparing
`obj.constructor` of the actual value, we started comparing the
constructor property as it exists on the prototype before the
constructor function runs, and thus also before any after-the-fact
mutation to an object property.

Undo that part of the change and add a regression test.

Fixes https://github.com/qunitjs/qunit/issues/1706.