Failing test: modules leak context to parents
Today, modules' `testEnvironment` is leaked to any parent module. This
makes it impossible for callers to rely on garbage collection of items
attached to the module context, which in turn makes it impossible to
test against memory leaks from the code under test. These failing tests
assert that the `this` of an inner module is not reference-equal to
the test of its parent module.