HTML Reporter: Move internal `urlparams` import out of html.js
As prep to extract a reporter-agnostic "HTML Runner" from the
HTML Reporter, move the `urlparams.js` import out of html.js.

It is important (and covered by `test/reporter-html/hidepassed.html`)
that the `QUnit.begin()` callback in urlparams.js continues to run
before the one in html.js.

When only moving the import statement, this breaks because diff.js
was also importing html.js for re-use of the `escapeText` function.
As such the order in reporter.js was not what it appeared.

It appeared as:

- fixture
- diff
- html
  - urlparams

But was actually:

- fixture
- diff
  - html
    - urlparams
- html (redundant)

Ref https://github.com/qunitjs/qunit/issues/1118