Test: add fixture for rejected "begin", "moduleStart", "done" callbacks
I'm adding this as a CLI fixture instead of an HTML fixture because
we don't currently have a good way of expecting a failure in an HTML
test run. I've confirmed that the "hanging" appearance described in
https://github.com/qunitjs/qunit/pull/1391 is equivalent to the
"Process exited before tests finished running" message reported.

Once we switch to capturing TAP by default, this will make the HTML
tests much easier to verify no matter whether the pass/fail outcome.

Ref https://github.com/qunitjs/qunit/pull/1391.