Build: Replace grunt-git-authors dependency with simple git-log command
I considered using git-shortlog as
`git shortlog -se --format='%aE' --no-merges`, which indeed gets us
fairly close. But, the available sort options there (as of git 2.41.0)
are by name or by count, not by first appearance.

So, instead, use git log, and pipe it to awk to deduplicate. This
is similar to piping to `sort | uniq`, except without the sorting.