Build: Update to ESLint 8 and Semistandard 17
In 9c3ae1a9be we updated to the latest supported versions for Node 10.
This commit updates to the latest supported for Node 18.

ESLint 9 requires a specific Node 18 release, and so can't be safely
selected without warnings from `ncu --enginesNode`. It requires either
ignoring that warning, or raising production requirements in package.json
to satisfy an internal dev dependency issue which is needlessly
inconveniences downstream users.