Core: Remove undocumented `details.modules[].tests` from QUnit.begin() event
This was added to QUnit.begin [1] in QUnit 1.16.0 for internal use by
html-reporter to know about each upcoming test before it is executed/finished.

In QUnit 2.7.0, we removed [2] that internal use in favour of listening to
the more reliable `QUnit.testStart()` [3] event instead since tests can sometimes
be defined later or during the test run, which would not be known to
`QUnit.begin().

[1]: https://qunitjs.com/api/callbacks/QUnit.begin/
[2]: https://github.com/qunitjs/qunit/commit/43a3d87fd1715a8229be27e5005a6883e69d760e
[3]: https://qunitjs.com/api/callbacks/QUnit.testStart/