HTML Reporter: Remove assertion count from qunit-testresult-display
QUnit 2.21.0:
> 254 tests completed in 912 milliseconds, with 0 failed, 7 skipped, and 4 todo.
> 819 assertions of 823 passed, 4 failed.

New:
> 254 tests completed in 912 milliseconds, with 0 failed, 7 skipped, and 4 todo.

For the reasoning behind this, see https://qunitjs.com/api/callbacks/QUnit.done/
and the deprecation of the `details` parameter to the QUnit.done() event.

In short, assertions are too low-level and include confusing/distracting mention
of failures that are not actually failures (but assertions in a todo test).
Plus it competes for attention when there is already a summary that is adequate
on its own.

The raw data remains available via `QUnit.config.stats.{all,bad}` for use in
integrations and plugins, although this is undocumented and not encouraged.
The recommendation is to use the `runEnd` and its stable test counts instead,
<https://qunitjs.com/api/callbacks/QUnit.on/#the-runend-event>.