HTML Reporter: Report time as seconds in qunit-testresult-display
QUnit 2.21.0:
> 254 tests completed in 912 milliseconds, with 0 failed, 7 skipped, and 4 todo.
> 819 assertions of 823 passed, 4 failed.

New:
> 254 tests completed in 0.9 seconds, with 0 failed, 7 skipped, and 4 todo.

This is a more human-scale number. Rounding is opionated. I went with 1 digit
of precision, and thus reporting 123ms as 0.1s. For number smaller than 100ms,
I opted for a longer format with at always 1 significant digit, so reporting
20ms as 0.02s, instead of e.g. forcefully rounding either down to a confusing
0.0s or a deceptively high 0.1s. It allows for a little pride/recognition of
small numbers. As a GUI, this is not meant to be machine readable either way.
In practice, numbers below 0.1s are rare.

The `runEnd` event data is unchanged, and continues to report in milliseconds.