Core: Update storage clear check to avoid QUnit.config.stats
Reduce internal reliance on the assertion count via the discouraged
QUnit.config.stats and its associated caveats as described at
https://qunitjs.com/api/callbacks/QUnit.done/.

This patch removes the last remaining internal use of `config.stats`.
After this we only ever set it, or expose it (in one place)
to provide the discouraged parameter to QUnit.done() callbacks for
back-compat.

Also remove outdated mention of internal config.stats.bad being used by
the HTML Reporter. This is no longer the case as of today with
https://github.com/qunitjs/qunit/pull/1760 landed for QUnit 3.0.