Build: Remove pushing of release artefact to detached Git commit
For QUnit 1.x and QUnit 2.x, we supported use of QUnit via
Git submodules, where we put the Git tag on a commit that isn't in the
main line, but instead a detached commit that contains the release
artefacts. This is also what Bower built on top of.

These two distribution methods will continue to work for old releases,
but QUnit 3.0 will instead only be canonically published to
code.jquery.com from where it can be downloaded over HTTP and HTTPS
(for setups that prefer something simpler than npm).

This means we don't need a separate "prepare" and "release" commit,
so merge these steps into one, and place the tag on the main branch
commit instead.

This also means we need to do our version field differently. We can:
* Keep "-pre" in Git, but have the release step temporarily remove it
  without including in the tagged commit.
* Change proper version for the release commit, then push another
  commit that changes it back.
* Use the proper version always, and let the build step inject "-dev"
  for non-release builds.

I went with the third option, which seems most conventional in the JS
ecosystem these days to simply leave it at the last release number
between releases.

Ref https://github.com/qunitjs/qunit/issues/1677.