HTML Reporter: Fix reversed order after clicking "Hide passed"
Follows-up https://github.com/qunitjs/qunit/pull/1311.

Since QUnit 2.7.0, when viewing a finished test run and turning on
"Hide passed" and then turning it off again, resulted in the same
results being shown again, but in reversed order.

This was because we hide the tests from top to bottom (and push into
the array), but then upon re-inserting them we used pop(), which means
we first append the last test to the end of the page, but then the
before-last is popped after that and appended to what is now the end
of the page, etc. The end result is the tests first displaying from
A-Z, then after toggling on/off, displaying from Z-A.