Core: Always define globalThis.QUnit
== Background ==

In QUnit 2.x, we always set the QUnit global in browser contexts, but
outside browsers, we only set the global if CJS/AMD wasn't detected.
Except in the QUnit CLI, where we do set the global anyway, because
that's how most QUnit tests are written, incliuding for Node.js.
So really the only case where the QUnit global is missing is custom
test runners that:

* run in Node.js,
* and require/import qunit.js directly,
* and don't export it as a global.

I'm aware a growing proportion of developers import 'qunit' directly
in each test file for improved type support. That's a great way
to avoid needing to rely on globals. But, it's not a requirement,
and is not always an option, especially for simple no-build-step
and browser-facing projects.

== Why ==

1. Improve portability between test runners.

Remove the last edge case where the QUnit global can be undefined.
Make it QUnit's responsiblity to define this reliably, as indeed
it almost always already does. Remove this as undocumented requirement
for specific test runners to patch up on their end.

In light of Karma deprecation, and emergence of more general purpose
TAP runners, I'd like to remove this as factor that might make/break
QUnit support in one of those.

2. Prepare for native ESM build.

We don't natively support ESM exports yet, but once we do this will
become a problem. To prevent split-brain problems with mixed use
(e.g. in test registry and other state) standardise internally on
which ever globalThis.QUnit was defined first, and then reliably
export that to any importers.

Ref https://github.com/qunitjs/qunit/issues/1551.