Browser: Finish extraction of browser runner from HTML Reporter
This refactors the initFixture and initUrlconfig code to be an exported
callable so that it is safe to define without side-effects and can then
be called conditionally.

The previous code was already conditionally with an inline check
for window/document. This is now centralised in prep for making it
further conditional on whether or not QUnit was already exported,
thus making it safe to load QUnit twice (e.g. ESM and CJS without
split-brain conflict). Tracked at
https://github.com/qunitjs/qunit/issues/1551.

Follows-up e1e03e68df788c013a7ccdc8ec1cc4aff33ec600.

Closes https://github.com/qunitjs/qunit/issues/1118.