Test: Refactor events.js and callbacks.js to regular fixtures
In particular, events.js was the only test we had in which a QUnit
test was loading another QUnit instance. This fails as-is after the
"Always define globalThis.QUnit" patch for
https://github.com/qunitjs/qunit/issues/1551, because it would return
the instance that's already running instead of a separate one.

It will still be possible to do this, by adding `delete global.QUnit`
before the import, but that's not the purpose of the events test.
Instead, I've added a separate test that proves inception works.