Docs: Fix expansion of "main" sidebar block on API Docs landing page
This restores something we briefly lost in the unification, since
api/index.md is now at "/api" rather than at "/" of a sub domain,
which is what `expand: initial` defaults to when deciding whether
to expand the sidebar block.

With this fix, we eliminates a click and potential point of confusion
when e.g. going from Home -> "View Docs" -> ? -> QUnit.test, by
removing the need to select a category.