HTML Reporter: Change `#qunit-testresult` from P to DIV
The previous DOM structure was impossible to represent in an HTML
string, which meant that systems like TestSwarm that serialize the
page and save or transfer it elsewhere, will end up with a corruption
when that same page is then retrieved and rendered again later, because
the element will be implicitly closed by the first DIV child.

The paragraph was already styled with `margin/padding: 0`, so this
makes no visual difference.

The Theme API has only been documented since yesterday during the
QUnit 3.0 dev cycle. Given this wasn't documented before, any change
is considered a breaking change and thus part of QUnit 3.0.

Going forward, the Theme API documents the selector for this by ID
only, not by tag name. That documentation was written based on how
most themes seem to style it, hence this is not likely to break
themes in practice.

Fixes https://github.com/qunitjs/qunit/issues/1301.