HTML Reporter: Change `#qunit-banner` from H2 to DIV
There is no need for this presentational pixel bar to be a heading
element. It is completely a presentational thing that is better
represented with a DIV instead.

This previously produced the following accessiblity warning,
e.g. in Firefox's Accessibility inspector, or in Chrome with AXE.

> Headings must not be labeled.
>
> `Role: heading. "" [text label]`
>
> https://developer.mozilla.org/en-US/docs/Web/Accessibility/Understanding_WCAG/Text_labels_and_names

> Headings must not be empty.
>
> https://github.com/dequelabs/axe-core/blob/v4.9.1/doc/rule-descriptions.md
> https://act-rules.github.io/rules/ffd0e9

Fixes https://github.com/qunitjs/qunit/issues/1427.